refactor: rename notable_id and notable_type to target_id and target_table in Notes model; reintroduce NotesMixin for polymorphic relationships
